Fixed multiple issues with the code.
