+++ app/Commands/InstallBreezeCommand.php
@@ -1,26 +1,31 @@
 <?php
 
-namespace DummyNamespace;
+namespace App\Commands;
 
 use Larakit\ComposerUsage;
 use Larakit\RunProcess;
 use LaravelZero\Framework\Commands\Command;
 
-class InstallDuskCommand extends Command
+class InstallBreezeCommand extends Command
 {
     /**
      * The signature of the command.
      *
      * @var string
      */
-    protected $signature = 'install:dusk';
+    protected $signature = 'install:breeze
+        {--dark : Adds dark mode support to Breeze}
+        {--vue : Adds Vue to Breeze}
+        {--react: Adds React to Breeze}
+        {--ssr: Adds Inertia SSR support to Breeze}
+        {--api: Scaffolds an authentication API for breeze}';
 
     /**
      * The description of the command.
      *
      * @var string
      */
-    protected $description = 'Installs Laravel Dusk to the Laravel application.';
+    protected $description = 'Installs Laravel Breeze to the Laravel application.';
 
     /**
      * Execute the console command.
@@ -29,7 +34,7 @@ class InstallDuskCommand extends Command
      */
     public function handle()
     {
-        $package = 'laravel/dusk';
+        $package = 'laravel/breeze';
 
         $dev = true;
 
@@ -49,7 +54,25 @@ class InstallDuskCommand extends Command
 
         $this->info($output);
 
-        $command = ['php', 'artisan', 'dusk:install'];
+        $command = ['php', 'artisan', 'breeze:install'];
+
+        if ($this->option('react')) {
+            $command[] = 'react';
+        }
+        if ($this->option('vue')) {
+            $command[] = 'vue';
+        }
+        if ($this->option('api')) {
+            $command[] = 'api';
+        }
+
+        if ($this->option('dark')) {
+            $command[] = '--dark';
+        }
+
+        if ($this->option('ssr')) {
+            $command[] = '--ssr';
+        }
 
         $this->info(RunProcess::run($command));
