+++ app/Commands/InstallFortifyCommand.php
@@ -1,26 +1,27 @@
 <?php
 
-namespace DummyNamespace;
+namespace App\Commands;
 
 use Larakit\ComposerUsage;
+use Larakit\RunProcess;
 use LaravelZero\Framework\Commands\Command;
 
-class InstallCashierCommand extends Command
+class InstallFortifyCommand extends Command
 {
     /**
      * The signature of the command.
      *
      * @var string
      */
-    protected $signature = 'install:cashier';
+    protected $signature = 'install:fortify
+        {--c|config : Installs the config file}';
 
     /**
      * The description of the command.
      *
      * @var string
      */
-    protected $description = 'Installs Laravel Cashier to the Laravel application.
-        {--p|paddle : Installs the Paddle version of Cashier}';
+    protected $description = 'Installs Laravel Fortify to the Laravel application.';
 
     /**
      * Execute the console command.
@@ -29,11 +30,7 @@ class InstallCashierCommand extends Command
      */
     public function handle()
     {
-        $package = 'laravel/cashier';
-
-        if ($this->option('paddle')) {
-            $package = $package.'-paddle';
-        }
+        $package = 'laravel/fortify';
 
         $dev = false;
 
@@ -54,6 +51,10 @@ class InstallCashierCommand extends Command
         $this->info($output);
 
         $this->comment($package.' installed successfully.');
+
+        if ($this->option('config')) {
+            $this->info(RunProcess::run(['php', 'artisan', 'vendor:publish', '--provider="Laravel\Fortify\FortifyServiceProvider"']));
+        }
 
         return Command::SUCCESS;
     }
