+++ stubs/installer/composer_installer.stub
@@ -1,6 +1,6 @@
 <?php
 
-namespace DummyNamespace;
+namespace App\Commands;
 
 use Illuminate\Console\Scheduling\Schedule;
 use LaravelZero\Framework\Commands\Command;
